Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 // Your code here...
 #include<stdio.h>
 int main(){
-    float n;
-    scanf("%f",&n);
+    double n;
+    scanf("%lf",&n);
     printf("You entered: %.4f",n);
     return 0;
 }
